+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+### Changed
+- Added coverage barrier to wf-human-variation
+
 ## [v1.3.0]
 ### Added
 - `--str` enables STR genotyping using Straglr (compatible with genome build 38 only)